refactor: remove Request.php trait and TableColumn.php trait
